PATCH: none

+++ Core/Source/Core/Types.h
@@ -9,9 +9,9 @@
 #include <fstream>
 
 #define SIZE(obj) sizeof(obj)/sizeof(obj[0])
-typedef u_int8_t u8;
-typedef u_int16_t u16;
-typedef u_int32_t u32;
+typedef uint8_t u8;
+typedef uint16_t u16;
+typedef uint32_t u32;
 typedef int8_t i8;
 typedef int16_t i16;
 typedef int32_t i32;
@@ -22,7 +22,5 @@ const u8 VIDEO_WIDTH = 64;
 const u8 VIDEO_HEIGHT = 32;
 
 // Initialize PC with a constructor
-const u8 FONTSET_START_ADDRESS = 0x50;
-
-
+const unsigned int FONTSET_START_ADDRESS = 0x50;
 #endif
